:sparkles: feat: MemberService 인터페이스 분리 #9
