+++ src/main/java/naeilmolae/domain/member/service/MemberService.java
@@ -1,22 +1,15 @@
 package naeilmolae.domain.member.service;
 
-import lombok.RequiredArgsConstructor;
 import naeilmolae.domain.member.domain.Member;
-import naeilmolae.domain.member.status.MemberErrorStatus;
-import naeilmolae.domain.member.repository.MemberRepository;
-import naeilmolae.global.common.exception.RestApiException;
+import naeilmolae.domain.member.dto.response.MemberIdResponseDto;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class MemberService {
-    private final MemberRepository memberRepository;
+public interface MemberService {
 
-    public Member findById(Long id) throws UsernameNotFoundException {
-        return memberRepository.findById(id)
-                .orElseThrow(() -> new RestApiException(MemberErrorStatus.EMPTY_MEMBER));
-    }
+    Member findById(Long id) throws UsernameNotFoundException;
+
+    // 회원 저장
+    Member saveEntity(Member member);
+    // 회원 탈퇴
+    MemberIdResponseDto withdrawal(Member member);
 }
