+++ src/main/java/naeilmolae/global/config/security/auth/PrincipalDetailsService.java
@@ -3,6 +3,7 @@ package naeilmolae.global.config.security.auth;
 import lombok.RequiredArgsConstructor;
 import naeilmolae.domain.member.domain.Member;
 import naeilmolae.domain.member.service.MemberService;
+import naeilmolae.domain.member.service.MemberServiceImpl;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
